backend/requirements.txt
fastapi
uvicorn
ultralytics
pillow
opencv-python
numpy